+++ test/test_doc_features.py
@@ -57,13 +57,13 @@ public:
     void array_method(int* ad, int size) {
         for (int i=0; i < size; ++i)
             std::cerr << ad[i] << ' ';
-        std::cerr << '\\n'; // TODO: not std::endl for 32b Windows
+        std::cerr << std::endl;
     }
 
     void array_method(double* ad, int size) {
         for (int i=0; i < size; ++i)
             std::cerr << ad[i] << ' ';
-        std::cerr << '\\n'; // TODO: not std::endl for 32b Windows
+        std::cerr << std::endl;
     }
 
     void uint_ref_assign(unsigned int& target, unsigned int value) {